chore: extract common transfer logic
