+++ sdk/src/main/java/com/hedera/hashgraph/sdk/TransferTransaction.java
@@ -21,36 +21,27 @@ package com.hedera.hashgraph.sdk;
 
 import com.google.common.base.MoreObjects;
 import com.google.protobuf.InvalidProtocolBufferException;
-import com.google.protobuf.UInt32Value;
 import com.hedera.hashgraph.sdk.proto.AccountAmount;
 import com.hedera.hashgraph.sdk.proto.CryptoServiceGrpc;
 import com.hedera.hashgraph.sdk.proto.CryptoTransferTransactionBody;
-import com.hedera.hashgraph.sdk.proto.NftTransfer;
 import com.hedera.hashgraph.sdk.proto.SchedulableTransactionBody;
 import com.hedera.hashgraph.sdk.proto.TransactionBody;
 import com.hedera.hashgraph.sdk.proto.TransactionResponse;
 import com.hedera.hashgraph.sdk.proto.TransferList;
 import io.grpc.MethodDescriptor;
-
-import javax.annotation.Nullable;
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 
 /**
- * A transaction that transfers hbars and tokens between Hedera accounts.
- * You can enter multiple transfers in a single transaction. The net value
- * of hbars between the sending accounts and receiving accounts must equal
- * zero.
- *
- * See <a href="https://docs.hedera.com/guides/docs/sdks/cryptocurrency/transfer-cryptocurrency">Hedera Documentation</a>
+ * A transaction that transfers hbars and tokens between Hedera accounts. You can enter multiple transfers in a single
+ * transaction. The net value of hbars between the sending accounts and receiving accounts must equal zero.
+ * <p>
+ * See <a href="https://docs.hedera.com/guides/docs/sdks/cryptocurrency/transfer-cryptocurrency">Hedera
+ * Documentation</a>
  */
-public class TransferTransaction extends Transaction<TransferTransaction> {
-    private final ArrayList<TokenTransfer> tokenTransfers = new ArrayList<>();
-    private final ArrayList<TokenNftTransfer> nftTransfers = new ArrayList<>();
+public class TransferTransaction extends AbstractTokenTransferTransaction<TransferTransaction> {
     private final ArrayList<HbarTransfer> hbarTransfers = new ArrayList<>();
 
     private static class HbarTransfer {
@@ -92,11 +83,12 @@ public class TransferTransaction extends Transaction<TransferTransaction> {
     /**
      * Constructor.
      *
-     * @param txs Compound list of transaction id's list of (AccountId, Transaction)
-     *            records
-     * @throws InvalidProtocolBufferException       when there is an issue with the protobuf
+     * @param txs Compound list of transaction id's list of (AccountId, Transaction) records
+     * @throws InvalidProtocolBufferException when there is an issue with the protobuf
      */
-    TransferTransaction(LinkedHashMap<TransactionId, LinkedHashMap<AccountId, com.hedera.hashgraph.sdk.proto.Transaction>> txs) throws InvalidProtocolBufferException {
+    TransferTransaction(
+        LinkedHashMap<TransactionId, LinkedHashMap<AccountId, com.hedera.hashgraph.sdk.proto.Transaction>> txs)
+        throws InvalidProtocolBufferException {
         super(txs);
         initFromTransactionBody();
     }
@@ -112,250 +104,9 @@ public class TransferTransaction extends Transaction<TransferTransaction> {
     }
 
     /**
-     * Extract the list of token id decimals.
-     *
-     * @return                          the list of token id decimals
-     */
-    public Map<TokenId, Integer> getTokenIdDecimals() {
-        Map<TokenId, Integer> decimalsMap = new HashMap<>();
-
-        for (var transfer : tokenTransfers) {
-            decimalsMap.put(transfer.tokenId, transfer.expectedDecimals);
-        }
-
-        return decimalsMap;
-    }
-
-    /**
-     * Extract the list of token transfer records.
-     *
-     * @return                          the list of token transfer records
-     */
-    public Map<TokenId, Map<AccountId, Long>> getTokenTransfers() {
-        Map<TokenId, Map<AccountId, Long>> transfers = new HashMap<>();
-
-        for (var transfer : tokenTransfers) {
-            var current = transfers.get(transfer.tokenId) != null
-                ? transfers.get(transfer.tokenId) : new HashMap<AccountId, Long>();
-            current.put(transfer.accountId, transfer.amount);
-            transfers.put(transfer.tokenId, current);
-        }
-
-        return transfers;
-    }
-
-    /**
-     * Add a token transfer to the transaction.
-     *
-     * @param tokenId                   the token id
-     * @param accountId                 the account id
-     * @param value                     the value
-     * @param isApproved                is it approved
-     * @return {@code this}
-     */
-    private TransferTransaction doAddTokenTransfer(TokenId tokenId, AccountId accountId, long value, boolean isApproved) {
-        requireNotFrozen();
-
-        for (var transfer : tokenTransfers) {
-            if (transfer.tokenId.equals(tokenId) && transfer.accountId.equals(accountId) && transfer.isApproved == isApproved) {
-                transfer.amount = transfer.amount + value;
-                return this;
-            }
-        }
-
-        tokenTransfers.add(new TokenTransfer(tokenId, accountId, value, isApproved));
-        return this;
-    }
-
-    /**
-     * Add a non approved token transfer to the transaction.
-     *
-     * @param tokenId                   the token id
-     * @param accountId                 the account id
-     * @param value                     the value
-     * @return                          the updated transaction
-     */
-    public TransferTransaction addTokenTransfer(TokenId tokenId, AccountId accountId, long value) {
-        return doAddTokenTransfer(tokenId, accountId, value, false);
-    }
-
-    /**
-     * Add an approved token transfer to the transaction.
-     *
-     * @param tokenId                   the token id
-     * @param accountId                 the account id
-     * @param value                     the value
-     * @return                          the updated transaction
-     */
-    public TransferTransaction addApprovedTokenTransfer(TokenId tokenId, AccountId accountId, long value) {
-        return doAddTokenTransfer(tokenId, accountId, value, true);
-    }
-
-    private TransferTransaction doAddTokenTransferWithDecimals(
-        TokenId tokenId,
-        AccountId accountId,
-        long value,
-        int decimals,
-        boolean isApproved
-    ) {
-        requireNotFrozen();
-
-        var found = false;
-
-        for (var transfer : tokenTransfers) {
-            if (transfer.tokenId.equals(tokenId)) {
-                if (transfer.expectedDecimals != null && transfer.expectedDecimals != decimals) {
-                    throw new IllegalArgumentException("expected decimals for a token in a token transfer cannot be changed after being set");
-                }
-
-                transfer.expectedDecimals = decimals;
-
-                if (transfer.accountId.equals(accountId) && transfer.isApproved == isApproved) {
-                    transfer.amount = transfer.amount + value;
-                    found = true;
-                }
-
-            }
-        }
-
-        if (found) {
-            return this;
-        }
-
-        tokenTransfers.add(new TokenTransfer(tokenId, accountId, value, decimals, isApproved));
-
-        return this;
-    }
-
-    /**
-     * Add a non approved token transfer with decimals.
-     *
-     * @param tokenId                   the token id
-     * @param accountId                 the account id
-     * @param value                     the value
-     * @param decimals                  the deco
-     * @return                          the updated transaction
-     */
-    public TransferTransaction addTokenTransferWithDecimals(
-        TokenId tokenId,
-        AccountId accountId,
-        long value,
-        int decimals
-    ) {
-        return doAddTokenTransferWithDecimals(tokenId, accountId, value, decimals, false);
-    }
-
-    /**
-     * Add an approved token transfer with decimals.
-     *
-     * @param tokenId                   the token id
-     * @param accountId                 the account id
-     * @param value                     the value
-     * @param decimals                  the deco
-     * @return                          the updated transaction
-     */
-    public TransferTransaction addApprovedTokenTransferWithDecimals(
-        TokenId tokenId,
-        AccountId accountId,
-        long value,
-        int decimals
-    ) {
-        return doAddTokenTransferWithDecimals(tokenId, accountId, value, decimals, true);
-    }
-
-    /**
-     * @deprecated - Use {@link #addApprovedTokenTransfer(TokenId, AccountId, long)} instead
-     * @param tokenId       the token id
-     * @param accountId     the account id
-     * @param isApproved    whether the transfer is approved
-     * @return {@code this}
-     */
-    @Deprecated
-    public TransferTransaction setTokenTransferApproval(TokenId tokenId, AccountId accountId, boolean isApproved) {
-        requireNotFrozen();
-
-        for (var transfer : tokenTransfers) {
-            if (transfer.tokenId.equals(tokenId) && transfer.accountId.equals(accountId)) {
-                transfer.isApproved = isApproved;
-                return this;
-            }
-        }
-
-        return this;
-    }
-
-    /**
-     * Extract the of token nft transfers.
-     *
-     * @return                          list of token nft transfers
-     */
-    public Map<TokenId, List<TokenNftTransfer>> getTokenNftTransfers() {
-        Map<TokenId, List<TokenNftTransfer>> transfers = new HashMap<>();
-
-        for (var transfer : nftTransfers) {
-            var current = transfers.get(transfer.tokenId) != null
-                ? transfers.get(transfer.tokenId) : new ArrayList<TokenNftTransfer>();
-            current.add(transfer);
-            transfers.put(transfer.tokenId, current);
-        }
-
-        return transfers;
-    }
-
-    private TransferTransaction doAddNftTransfer(NftId nftId, AccountId sender, AccountId receiver, boolean isApproved) {
-        requireNotFrozen();
-        nftTransfers.add(new TokenNftTransfer(nftId.tokenId, sender, receiver, nftId.serial, isApproved));
-        return this;
-    }
-
-    /**
-     * Add a non approved nft transfer.
-     *
-     * @param nftId                     the nft's id
-     * @param sender                    the sender account id
-     * @param receiver                  the receiver account id
-     * @return                          the updated transaction
-     */
-    public TransferTransaction addNftTransfer(NftId nftId, AccountId sender, AccountId receiver) {
-        return doAddNftTransfer(nftId, sender, receiver, false);
-    }
-
-    /**
-     * Add an approved nft transfer.
-     *
-     * @param nftId                     the nft's id
-     * @param sender                    the sender account id
-     * @param receiver                  the receiver account id
-     * @return                          the updated transaction
-     */
-    public TransferTransaction addApprovedNftTransfer(NftId nftId, AccountId sender, AccountId receiver) {
-        return doAddNftTransfer(nftId, sender, receiver, true);
-    }
-
-    /**
-     * @deprecated - Use {@link #addApprovedNftTransfer(NftId, AccountId, AccountId)} instead
-     * @param nftId         the NFT id
-     * @param isApproved    whether the transfer is approved
-     * @return {@code this}
-     */
-    @Deprecated
-    public TransferTransaction setNftTransferApproval(NftId nftId, boolean isApproved) {
-        requireNotFrozen();
-
-        for (var transfer : nftTransfers) {
-            if (transfer.tokenId.equals(nftId.tokenId) && transfer.serial == nftId.serial) {
-                transfer.isApproved = isApproved;
-                return this;
-            }
-        }
-
-        return this;
-    }
-
-    /**
      * Extract the of hbar transfers.
      *
-     * @return                          list of hbar transfers
+     * @return list of hbar transfers
      */
     public Map<AccountId, Hbar> getHbarTransfers() {
         Map<AccountId, Hbar> transfers = new HashMap<>();
@@ -384,9 +135,9 @@ public class TransferTransaction extends Transaction<TransferTransaction> {
     /**
      * Add a non approved hbar transfer to an EVM address.
      *
-     * @param evmAddress                the EVM address
-     * @param value                     the value
-     * @return                          the updated transaction
+     * @param evmAddress the EVM address
+     * @param value      the value
+     * @return the updated transaction
      */
     public TransferTransaction addHbarTransfer(EvmAddress evmAddress, Hbar value) {
         AccountId accountId = AccountId.fromEvmAddress(evmAddress);
@@ -396,9 +147,9 @@ public class TransferTransaction extends Transaction<TransferTransaction> {
     /**
      * Add a non approved hbar transfer.
      *
-     * @param accountId                 the account id
-     * @param value                     the value
-     * @return                          the updated transaction
+     * @param accountId the account id
+     * @param value     the value
+     * @return the updated transaction
      */
     public TransferTransaction addHbarTransfer(AccountId accountId, Hbar value) {
         return doAddHbarTransfer(accountId, value, false);
@@ -407,19 +158,19 @@ public class TransferTransaction extends Transaction<TransferTransaction> {
     /**
      * Add an approved hbar transfer.
      *
-     * @param accountId                 the account id
-     * @param value                     the value
-     * @return                          the updated transaction
+     * @param accountId the account id
+     * @param value     the value
+     * @return the updated transaction
      */
     public TransferTransaction addApprovedHbarTransfer(AccountId accountId, Hbar value) {
         return doAddHbarTransfer(accountId, value, true);
     }
 
     /**
-     * @deprecated - Use {@link #addApprovedHbarTransfer(AccountId, Hbar)} instead
-     * @param accountId     the account id
-     * @param isApproved    whether the transfer is approved
+     * @param accountId  the account id
+     * @param isApproved whether the transfer is approved
      * @return {@code this}
+     * @deprecated - Use {@link #addApprovedHbarTransfer(AccountId, Hbar)} instead
      */
     @Deprecated
     public TransferTransaction setHbarTransferApproval(AccountId accountId, boolean isApproved) {
@@ -438,80 +189,20 @@ public class TransferTransaction extends Transaction<TransferTransaction> {
     /**
      * Build the transaction body.
      *
-     * @return {@link
-     *         com.hedera.hashgraph.sdk.proto.CryptoTransferTransactionBody}
+     * @return {@link com.hedera.hashgraph.sdk.proto.CryptoTransferTransactionBody}
      */
     CryptoTransferTransactionBody.Builder build() {
-        var tokenTransfers = new ArrayList<TokenTransferList>();
-
-        this.hbarTransfers.sort(Comparator.comparing((HbarTransfer a) -> a.accountId).thenComparing(a -> a.isApproved));
-        this.tokenTransfers.sort(Comparator.comparing((TokenTransfer a) -> a.tokenId).thenComparing(a -> a.accountId).thenComparing(a -> a.isApproved));
-        this.nftTransfers.sort(Comparator.comparing((TokenNftTransfer a) -> a.tokenId).thenComparing(a -> a.sender).thenComparing(a -> a.receiver).thenComparing(a -> a.serial));
-
-        var i = 0;
-        var j = 0;
-
-        // Effectively merge sort
-        while (i < this.tokenTransfers.size() || j < this.nftTransfers.size()) {
-            if (i < this.tokenTransfers.size() && j < this.nftTransfers.size()) {
-                var iTokenId = this.tokenTransfers.get(i).tokenId;
-                var jTokenId = this.nftTransfers.get(j).tokenId;
-                var last = !tokenTransfers.isEmpty() ? tokenTransfers.get(tokenTransfers.size() - 1) : null;
-                var lastTokenId = last != null ? last.tokenId : null;
-
-                if (last != null && iTokenId.compareTo(lastTokenId) == 0) {
-                    last.transfers.add(this.tokenTransfers.get(i++));
-                    continue;
-                }
-
-                if (last != null && jTokenId.compareTo(lastTokenId) == 0) {
-                    last.nftTransfers.add(this.nftTransfers.get(j++));
-                    continue;
-                }
-
-                var result = iTokenId.compareTo(jTokenId);
-
-                if (result == 0) {
-                    tokenTransfers.add(new TokenTransferList(iTokenId, this.tokenTransfers.get(i).expectedDecimals, this.tokenTransfers.get(i++), this.nftTransfers.get(j++)));
-                } else if (result < 0) {
-                    tokenTransfers.add(new TokenTransferList(iTokenId, this.tokenTransfers.get(i).expectedDecimals, this.tokenTransfers.get(i++), null));
-                } else {
-                    tokenTransfers.add(new TokenTransferList(jTokenId, null, null, this.nftTransfers.get(j++)));
-                }
-            } else if (i < this.tokenTransfers.size()) {
-                var iTokenId = this.tokenTransfers.get(i).tokenId;
-                var last = !tokenTransfers.isEmpty() ? tokenTransfers.get(tokenTransfers.size() - 1) : null;
-                var lastTokenId = last != null ? last.tokenId : null;
-
-                if (last != null && iTokenId.compareTo(lastTokenId) == 0) {
-                    last.transfers.add(this.tokenTransfers.get(i++));
-                    continue;
-                }
-
-                tokenTransfers.add(new TokenTransferList(iTokenId, this.tokenTransfers.get(i).expectedDecimals, this.tokenTransfers.get(i++), null));
-            } else {
-                var jTokenId = this.nftTransfers.get(j).tokenId;
-                var last = !tokenTransfers.isEmpty() ? tokenTransfers.get(tokenTransfers.size() - 1) : null;
-                var lastTokenId = last != null ? last.tokenId : null;
-
-                if (last != null && jTokenId.compareTo(lastTokenId) == 0) {
-                    last.nftTransfers.add(this.nftTransfers.get(j++));
-                    continue;
-                }
-
-                tokenTransfers.add(new TokenTransferList(jTokenId, null, null, this.nftTransfers.get(j++)));
-            }
-        }
+        var transfers = sortTransfersAndBuild();
 
         var builder = CryptoTransferTransactionBody.newBuilder();
 
-        var transfers = TransferList.newBuilder();
+        var hbarTransfersList = TransferList.newBuilder();
         for (var transfer : hbarTransfers) {
-            transfers.addAccountAmounts(transfer.toProtobuf());
+            hbarTransfersList.addAccountAmounts(transfer.toProtobuf());
         }
-        builder.setTransfers(transfers);
+        builder.setTransfers(hbarTransfersList);
 
-        for (var transfer : tokenTransfers) {
+        for (var transfer : transfers) {
             builder.addTokenTransfers(transfer.toProtobuf());
         }
 
@@ -520,18 +211,8 @@ public class TransferTransaction extends Transaction<TransferTransaction> {
 
     @Override
     void validateChecksums(Client client) throws BadEntityIdException {
+        super.validateChecksums(client);
         for (var transfer : hbarTransfers) {
-            transfer.accountId.validateChecksum(client);
-        }
-
-        for (var transfer : nftTransfers) {
-            transfer.tokenId.validateChecksum(client);
-            transfer.sender.validateChecksum(client);
-            transfer.receiver.validateChecksum(client);
-        }
-
-        for (var transfer : tokenTransfers) {
-            transfer.tokenId.validateChecksum(client);
             transfer.accountId.validateChecksum(client);
         }
     }
